Update to use trigger apply for creation and updating
