+++ internal/cli/trigger_apply.go
@@ -3,12 +3,37 @@ package cli
 import (
 	"github.com/posener/complete"
 
+	"github.com/hashicorp/waypoint-plugin-sdk/terminal"
+	"github.com/hashicorp/waypoint/internal/clierrors"
 	"github.com/hashicorp/waypoint/internal/pkg/flag"
+	pb "github.com/hashicorp/waypoint/internal/server/gen"
 )
 
 type TriggerApplyCommand struct {
 	*baseCommand
+
+	flagTriggerName        string
+	flagTriggerId          string
+	flagTriggerDescription string
+	flagTriggerLabels      []string
+	flagTriggerOperation   string
+	flagTriggerNoAuth      bool
+
+	// Operation options
+	flagArtifactSeq int
+	flagBuildSeq    int
+	flagDeploySeq   int
+	flagReleaseSeq  int
+	flagDisablePush bool
+
+	// Release options
+	flagReleasePrune       bool
+	flagReleasePruneRetain int
 }
+
+// Current supported trigger operation names.
+var triggerOpValues = []string{"build", "push", "deploy", "destroy-workspace",
+	"destroy-deployment", "release", "up", "init"}
 
 func (c *TriggerApplyCommand) Run(args []string) int {
 	// Initialize. If we fail, we just exit since Init handles the UI.
@@ -20,12 +45,287 @@ func (c *TriggerApplyCommand) Run(args []string) int {
 	); err != nil {
 		return 1
 	}
+	ctx := c.Ctx
+
+	var diffTrigger *pb.Trigger
+	if c.flagTriggerId != "" {
+		// Look for an existing trigger if id specified
+		respTrigger, err := c.project.Client().GetTrigger(ctx, &pb.GetTriggerRequest{
+			Ref: &pb.Ref_Trigger{
+				Id: c.flagTriggerId,
+			},
+		})
+		if err != nil {
+			c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
+			return 1
+		}
+
+		diffTrigger = respTrigger.Trigger
+	}
+
+	// for now we don't let people change the trigger operation or what the target is
+	if diffTrigger != nil {
+		if c.flagTriggerName == "" {
+			c.flagTriggerName = diffTrigger.Name
+		}
+		if c.flagTriggerDescription == "" {
+			c.flagTriggerDescription = diffTrigger.Description
+		}
+		if len(c.flagTriggerLabels) == 0 {
+			c.flagTriggerLabels = diffTrigger.Labels
+		}
+	}
+
+	createTrigger := &pb.Trigger{
+		Name:          c.flagTriggerName,
+		Description:   c.flagTriggerDescription,
+		Labels:        c.flagTriggerLabels,
+		Authenticated: !c.flagTriggerNoAuth,
+		Workspace: &pb.Ref_Workspace{
+			Workspace: c.flagWorkspace,
+		},
+		Project: &pb.Ref_Project{
+			Project: c.flagProject,
+		},
+		Application: &pb.Ref_Application{
+			Application: c.flagApp,
+			Project:     c.flagProject,
+		},
+	}
+
+	if diffTrigger != nil {
+		createTrigger.Id = diffTrigger.Id
+	}
+
+	// Set the operation
+	switch {
+	case c.flagTriggerOperation == "build":
+		createTrigger.Operation = &pb.Trigger_Build{
+			Build: &pb.Job_BuildOp{
+				DisablePush: c.flagDisablePush,
+			},
+		}
+	case c.flagTriggerOperation == "push":
+		if c.flagBuildSeq == 0 {
+			c.ui.Output("Must specify a build sequence number for the \"push\" operation: %s",
+				c.Flags().Help(), terminal.WithErrorStyle())
+			return 1
+		}
+
+		// TODO: Check ws, proj, app too? maybe not, API could check it
+
+		createTrigger.Operation = &pb.Trigger_Push{
+			Push: &pb.Job_PushOp{
+				Build: &pb.Build{
+					Sequence: uint64(c.flagBuildSeq),
+					Workspace: &pb.Ref_Workspace{
+						Workspace: c.flagWorkspace,
+					},
+					Application: &pb.Ref_Application{
+						Application: c.flagApp,
+						Project:     c.flagProject,
+					},
+				},
+			},
+		}
+	case c.flagTriggerOperation == "deploy":
+		// TODO/NOTE: If no sequence number is specififed (i.e. seq 0), the backend should default to using the "latest" artifact instead
+
+		createTrigger.Operation = &pb.Trigger_Deploy{
+			Deploy: &pb.Job_DeployOp{
+				Artifact: &pb.PushedArtifact{
+					Sequence: uint64(c.flagBuildSeq),
+					Workspace: &pb.Ref_Workspace{
+						Workspace: c.flagWorkspace,
+					},
+					Application: &pb.Ref_Application{
+						Application: c.flagApp,
+						Project:     c.flagProject,
+					},
+				},
+			},
+		}
+	case c.flagTriggerOperation == "destroy-workspace":
+		// NOTE(briancain): I don't think this operation actually works, takes no arguments...
+		createTrigger.Operation = &pb.Trigger_Destroy{
+			Destroy: &pb.Job_DestroyOp{
+				Target: &pb.Job_DestroyOp_Workspace{},
+			},
+		}
+	case c.flagTriggerOperation == "destroy-deployment":
+		createTrigger.Operation = &pb.Trigger_Destroy{
+			Destroy: &pb.Job_DestroyOp{
+				Target: &pb.Job_DestroyOp_Deployment{
+					Deployment: &pb.Deployment{
+						Sequence: uint64(c.flagDeploySeq),
+						Workspace: &pb.Ref_Workspace{
+							Workspace: c.flagWorkspace,
+						},
+						Application: &pb.Ref_Application{
+							Application: c.flagApp,
+							Project:     c.flagProject,
+						},
+					},
+				},
+			},
+		}
+	case c.flagTriggerOperation == "release":
+		// if no deployment seq is specified, the backend should default to latest
+		rt := &pb.Trigger_Release{
+			Release: &pb.Job_ReleaseOp{
+				Deployment: &pb.Deployment{
+					Sequence: uint64(c.flagDeploySeq),
+					Workspace: &pb.Ref_Workspace{
+						Workspace: c.flagWorkspace,
+					},
+					Application: &pb.Ref_Application{
+						Application: c.flagApp,
+						Project:     c.flagProject,
+					},
+				},
+				Prune: c.flagReleasePrune,
+			},
+		}
+
+		if c.flagReleasePruneRetain > 0 {
+			rt.Release.PruneRetain = int32(c.flagReleasePruneRetain)
+			rt.Release.PruneRetainOverride = true
+		}
+
+		createTrigger.Operation = rt
+	case c.flagTriggerOperation == "up":
+		releaseOp := &pb.Job_ReleaseOp{
+			Prune: c.flagReleasePrune,
+		}
+
+		if c.flagReleasePruneRetain > 0 {
+			releaseOp.PruneRetain = int32(c.flagReleasePruneRetain)
+			releaseOp.PruneRetainOverride = true
+		}
+
+		createTrigger.Operation = &pb.Trigger_Up{
+			Up: &pb.Job_UpOp{
+				Release: releaseOp,
+			},
+		}
+	case c.flagTriggerOperation == "init":
+		createTrigger.Operation = &pb.Trigger_Init{
+			Init: &pb.Job_InitOp{},
+		}
+	default:
+		// This shouldn't happened because the flag package should technically be handling the parsing
+		// and fail if any value was not recognized in the defined Enum
+		c.ui.Output("Unrecognized operation type %q: ", c.flagTriggerOperation, terminal.WithErrorStyle())
+		return 1
+	}
+
+	resp, err := c.project.Client().UpsertTrigger(ctx, &pb.UpsertTriggerRequest{
+		Trigger: createTrigger,
+	})
+	if err != nil {
+		c.ui.Output(clierrors.Humanize(err), terminal.WithErrorStyle())
+		return 1
+	}
+
+	// TODO: update output to show trigger URL with wp server attached
+	c.ui.Output("Trigger %q (%s) has been created", resp.Trigger.Name, resp.Trigger.Id, terminal.WithSuccessStyle())
 
 	return 0
 }
 
 func (c *TriggerApplyCommand) Flags() *flag.Sets {
-	return c.flagSet(flagSetOperation, nil)
+	return c.flagSet(flagSetOperation, func(set *flag.Sets) {
+		f := set.NewSet("Command Options")
+
+		f.StringVar(&flag.StringVar{
+			Name:    "name",
+			Target:  &c.flagTriggerName,
+			Default: "",
+			Usage:   "The name the trigger configuration should be defined as.",
+		})
+
+		f.StringVar(&flag.StringVar{
+			Name:   "id",
+			Target: &c.flagTriggerId,
+			Usage: "If specified, will look up an existing trigger by this id and " +
+				"attempt to update the configuration.",
+		})
+
+		f.StringVar(&flag.StringVar{
+			Name:    "description",
+			Target:  &c.flagTriggerDescription,
+			Default: "",
+			Usage:   "A human readable description about the trigger URL configuration.",
+		})
+
+		f.StringSliceVar(&flag.StringSliceVar{
+			Name:   "trigger-label",
+			Target: &c.flagTriggerLabels,
+			Usage:  "A collection of labels to apply to the trigger URL configuration. Can be specified multiple times.",
+		})
+
+		f.BoolVar(&flag.BoolVar{
+			Name:    "no-auth",
+			Target:  &c.flagTriggerNoAuth,
+			Default: false,
+			Usage:   "If set, the trigger URL configuration will not require authentication to initiate a request.",
+		})
+
+		f.EnumSingleVar(&flag.EnumSingleVar{
+			Name:   "op",
+			Target: &c.flagTriggerOperation,
+			Values: triggerOpValues,
+			Usage:  "The operation the trigger should execute when requested.",
+		})
+
+		// Operation specific flags
+		fo := set.NewSet("Operation Options")
+		fo.IntVar(&flag.IntVar{
+			Name:   "artifact-sequence",
+			Target: &c.flagArtifactSeq,
+			Usage:  "The sequence number for the artifact to use in an operation.",
+		})
+
+		fo.BoolVar(&flag.BoolVar{
+			Name:    "disable-push",
+			Target:  &c.flagDisablePush,
+			Default: false,
+			Usage:   "Disables pushing a build artifact to any configured registry for build operations.",
+		})
+
+		fo.IntVar(&flag.IntVar{
+			Name:   "build-sequence",
+			Target: &c.flagBuildSeq,
+			Usage:  "The sequence number for the build to use in an operation.",
+		})
+
+		fo.IntVar(&flag.IntVar{
+			Name:   "deployment-sequence",
+			Target: &c.flagDeploySeq,
+			Usage:  "The sequence number for the deployment to use in an operation.",
+		})
+
+		fo.IntVar(&flag.IntVar{
+			Name:   "release-sequence",
+			Target: &c.flagReleaseSeq,
+			Usage:  "The sequence number for the release to use in an operation.",
+		})
+
+		// Release operation specific flags
+		fro := set.NewSet("Release Operation Options")
+		fro.BoolVar(&flag.BoolVar{
+			Name:    "prune",
+			Target:  &c.flagReleasePrune,
+			Default: false,
+			Usage:   "If true, will prune deployments that aren't released.",
+		})
+
+		fro.IntVar(&flag.IntVar{
+			Name:   "prune-retain",
+			Target: &c.flagReleasePruneRetain,
+			Usage:  "This sets the number of unreleased deployments to retain when pruning.",
+		})
+	})
 }
 
 func (c *TriggerApplyCommand) AutocompleteArgs() complete.Predictor {
@@ -37,14 +337,18 @@ func (c *TriggerApplyCommand) AutocompleteFlags() complete.Flags {
 }
 
 func (c *TriggerApplyCommand) Synopsis() string {
-	return "Update a trigger URL configuration on Waypoint server"
+	return "Generate and Update a trigger URL and register it to Waypoint server"
 }
 
 func (c *TriggerApplyCommand) Help() string {
 	return formatHelp(`
 Usage: waypoint trigger apply [options]
 
-  Update a trigger URL configuration on Waypoint Server.
+  Create or update a trigger URL to Waypoint Server.
+
+	If no sequence number is specified, the trigger will use the "latest" sequence
+	for the given operation. I.e. if you create a deploy trigger with no specified
+	build artifact sequence number, it will use whatever the latest artifact sequence is.
 
 ` + c.Flags().Help())
 }
